move toc utils to theme-common
